+++ Shared/Models/Products/Product.cs
@@ -25,11 +25,11 @@ public class Product
     [Required(ErrorMessage = "Sell Price is required")]
     public decimal? SellPrice { get; set; }
 
-    public int DispensaryQuantity => Dispensary.Where(x => x.Quantity > 0).Sum(x => x.Quantity).GetValueOrDefault();
-    public int QuantitySold => OrderItems.Where(x => x.Status == OrderStatus.Completed).Sum(x => x.Quantity);    
-    public int QuantityPending => OrderItems.Where(x => x.Status == OrderStatus.Pending).Sum(x => x.Quantity);    
-    public int QuantityCancelled => OrderItems.Where(x => x.Status == OrderStatus.Canceled).Sum(x => x.Quantity);    
-    public int StoreQuantity => Stocks.Where(x => x.Quantity > 0).Sum(x => x.Quantity).GetValueOrDefault();
+    public decimal DispensaryQuantity => Dispensary.Where(x => x.Quantity > 0).Sum(x => x.Quantity).GetValueOrDefault();
+    public decimal QuantitySold => OrderItems.Where(x => x.Status == OrderStatus.Completed).Sum(x => x.Quantity);    
+    public decimal QuantityPending => OrderItems.Where(x => x.Status == OrderStatus.Pending).Sum(x => x.Quantity);    
+    public decimal QuantityCancelled => OrderItems.Where(x => x.Status == OrderStatus.Canceled).Sum(x => x.Quantity);    
+    public decimal StoreQuantity => Stocks.Where(x => x.Quantity > 0).Sum(x => x.Quantity).GetValueOrDefault();
     public int ReorderLevel { get; set; }
 
     public DateTime CreatedDate { get; set; } = DateTime.Now;
