+++ Shared/Models/Products/RestockingModel.cs
@@ -11,8 +11,8 @@ public class RestockingModel
     public string? Option { get; set; }
     public Guid ProductId { get; set; }
     public string? ProductName { get; set; }    
-    public int CurrentQty { get; set; }
-    public int? NewQty { get; set; }
+    public decimal CurrentQty { get; set; }
+    public decimal? NewQty { get; set; }
     public Guid StoreID { get; set; }
     public Guid StockID { get; set; }
     public DateTime? ExpiryDate { get; set; }
